+++ src/components/Parser.js
@@ -14,13 +14,24 @@ class Parser extends React.Component {
 	}
 
 	render() {
-		let inner_class   = (this.props.collapsed && this.props.viewing === "tables") ? "auto-height" : "";
-		let opacity       = this.props.opacity / 100;
-		let opacity_style = `
+		let collapsed          = (this.props.collapsed && this.props.viewing === "tables");
+		let root_inner_classes = [];
+		let opacity            = this.props.opacity / 100;
+		let opacity_style      = `
 			#container {
 				opacity: ${opacity};
 			}
 		`;
+
+		if (collapsed) {
+			root_inner_classes.push("auto-height");
+
+			if (this.props.collapse_down) {
+				root_inner_classes.push("down");
+			}
+		}
+
+		root_inner_classes = root_inner_classes.join(" ");
 
 		return (
 			<React.Fragment>
@@ -28,7 +39,7 @@ class Parser extends React.Component {
 					{opacity_style}
 					{this.props.css}
 				</style>
-				<div id="root-inner" className={inner_class}>
+				<div id="root-inner" className={root_inner_classes}>
 					<Container/>
 				</div>
 			</React.Fragment>
@@ -38,10 +49,11 @@ class Parser extends React.Component {
 
 const mapStateToProps = (state) => {
 	return {
-		collapsed : state.settings.intrinsic.collapsed,
-		viewing   : state.internal.viewing,
-		css       : state.settings.custom.css || "",
-		opacity   : state.settings.interface.opacity
+		collapsed     : state.settings.intrinsic.collapsed,
+		collapse_down : state.settings.interface.collapse_down,
+		viewing       : state.internal.viewing,
+		css           : state.settings.custom.css || "",
+		opacity       : state.settings.interface.opacity
 	};
 };